pyproject.toml
[project]
name = "vtok"
version = "0.1.0"
description = "Unofficial implementation of VTok: A Unified Video Tokenizer with Decoupled Spatial-Temporal Latents"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2",
    "torchvision>=0.17",
    "transformers>=4.40",
    "diffusers>=0.27",
    "pyyaml>=6.0",
    "wandb>=0.18",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.7",
]

[project.scripts]
vtok-train = "vtok.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["."]
include = ["vtok*"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]